requirements.txt
flask
werkzeug.security
pytest
flask_sqlalchemy
fastjson-db==0.3.5
python-dotenv